+++ device_calendar/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 3.1.0
+
+* Boolean variable `isDefault` added for issue [145](https://github.com/builttoroam/flutter_plugins/issues/145) (**NOTE**: This is not supported Android API 16 or lower, `isDefault` will always be false)
+* Events with 'null' title now defaults to 'New Event', issue [126](https://github.com/builttoroam/flutter_plugins/issues/126)
+* Updated property summaries for issues [121](https://github.com/builttoroam/flutter_plugins/issues/121) and [122](https://github.com/builttoroam/flutter_plugins/issues/122)
+* Updated example documentation for issue [119](https://github.com/builttoroam/flutter_plugins/issues/119)
+
 ## 3.0.0 21st January 2020
 
 * **BREAKING CHANGE** Properties for the attendee model in `attendee.dart` file have been changed: